% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bullying_data}
\alias{bullying_data}
\title{Bullying response data}
\format{
## `bullying_data`
A list of length 414:
\describe{
  \item{bullying_data}{response text}
}
}
\source{
<https://urn.fi/urn:nbn:fi:fsd:T-FSD3134>
}
\usage{
bullying_data
}
\description{
This data contains the responses to q7 "Kertoisitko, mitä sinun mielestäsi
kiusaaminen on? (Avokysymys)" in the FSD3134 Lapsibarometri 2016 dataset.
}
\keyword{datasets}
