% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{child_barometer_data}
\alias{child_barometer_data}
\title{Child Barometer 2016 response data}
\format{
## `child_barometer_data`
A dataframe with 414 rows and 2 columns:
\describe{
  \item{fsd_id}{FSD case id}
  \item{q7}{response text}
}
}
\source{
<https://urn.fi/urn:nbn:fi:fsd:T-FSD3134>
}
\usage{
child_barometer_data
}
\description{
This data contains the responses to q7 "Kertoisitko, mitä sinun mielestäsi
kiusaaminen on? (Avokysymys)" in the FSD3134 Lapsibarometri 2016 dataset.
}
\keyword{datasets}
