% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dev_data_f}
\alias{dev_data_f}
\title{Young People's Views on Development Cooperation 2012 Female response data}
\format{
## `dev_data_f`
A dataframe with 673 rows and 4 columns:
\describe{
  \item{fsd_id}{FSD case id}
  \item{q11_1}{response text for q11_1}
  \item{q11_2}{response text for q11_2}
  \item{q11_3}{response text for q11_3}
}
}
\source{
<https://urn.fi/urn:nbn:fi:fsd:T-FSD2821>
}
\usage{
dev_data_f
}
\description{
This data contains the female responses to q11_1 'Jatka lausetta: Kehitysmaa
on maa, jossa... (Avokysymys)', q11_2 'Jatka lausetta: Kehitysyhteistyö on
toimintaa, jossa... (Avokysymys)', q11_3' Jatka lausetta: Maailman kolme
suurinta ongelmaa ovat... (Avokysymys)' in the FSD2821 Nuorten ajatuksia
kehitysyhteistyöstä 2012 dataset.
}
\keyword{datasets}
