% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{english_sample_survey}
\alias{english_sample_survey}
\title{English Sample Survey Data: Patient Joe}
\format{
## `english_sample_survey`
A dataframe with 585 rows and 5 columns:
\describe{
  \item{id}{ID}
  \item{label}{Label: destructive, passive, somewhat proactive, or proactive}
  \item{label_coder1}{Label from coder 1}
  \item{label_coder2}{Label from coder 2}
  \item{text}{Text of response}
}
}
\source{
<https://doi.org/10.7802/2474>
}
\usage{
english_sample_survey
}
\description{
This data contains English text responses to ""Joe’s doctor told him that he
would need to return in two weeks to find out whether or not his condition
had improved. But when Joe asked the receptionist for an appointment, he was
told that it would be over a month before the next available appointment.
What should Joe do?" as well as categorisation of these responses by two
coders as either destructive, passive, somewhat proactive, or proactive.
}
\keyword{datasets}
