% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_concept_network.R
\name{fst_cn_nodes}
\alias{fst_cn_nodes}
\title{Concept Network - Get Textrank Nodes}
\usage{
fst_cn_nodes(data, edges, relevant_pos = c("NOUN", "VERB", "ADJ", "ADV"))
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}

\item{edges}{Output of fst_cn_edges, dataframe of 'edges' connecting two words}

\item{relevant_pos}{List of UPOS tags for inclusion, default is c("NOUN",
"VERB", "ADJ", "ADV").}
}
\value{
A dataframe containing relevant lemmas and their associated pagerank
}
\description{
Concept Network - Get Textrank Nodes
}
\examples{
lonely_nodes_2 <- fst_cn_nodes(conllu_lonely, lonely_edges_2)
lonely_nodes <- fst_cn_nodes(conllu_lonely_nltk, lonely_edges)
bullying_nodes <- fst_cn_nodes(conllu_bullying, bullying_edges)
}
