% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_concept_network.R
\name{fst_cn_plot}
\alias{fst_cn_plot}
\title{Plot Concept Network}
\usage{
fst_cn_plot(edges, nodes, concepts)
}
\arguments{
\item{edges}{Output of `fst_cn_edges`, dataframe of 'edges' connecting two words}

\item{nodes}{Output of `fst_cn_nodes`, dataframe of relevant lemmas and their associated pagerank}

\item{concepts}{List of terms which have been searched for, separated by commas.}
}
\value{
Plot of Concept Network
}
\description{
Plot Concept Network
}
\examples{
fst_cn_plot(edges = lonely_edges_2, nodes = lonely_nodes_2, concepts = 'yksinäisyys, tunne, tuntea')
fst_cn_plot(edges = lonely_edges, nodes = lonely_nodes, concepts = 'yksinäisyys, tunne, tuntea')
fst_cn_plot(edges = bullying_edges, nodes = bullying_nodes, concepts = 'kiusata, lyöminen')
}
