% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_discover}
\alias{fst_discover}
\title{Perform Data Exploration}
\usage{
fst_discover(
  data,
  freq_number = 20,
  ngram_number = 20,
  ngrams = 2,
  pos_filter = NULL
)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}

\item{freq_number}{The number of top words to return, default is `20`.}

\item{ngram_number}{The number of top n-grams to return, default is `20`.}

\item{ngrams}{The type of n-grams to return, default is `2`.}

\item{pos_filter}{List of UPOS tags for inclusion, default is `NULL` which
means all word types included.}
}
\value{
All the plots from the data exploration functions in the plots pane
}
\description{
This summary function runs all the data exploration functions -
`fst_summarise`, `fst_pos`, `fst_freq`, `fst_ngrams`, and `fst_wordcloud`
within one step. These can be viewed within the `plots´ pane of RStudio by
using the left and right arrows to cycle between.
}
\examples{
fst_discover(bullying)
fst_discover(data, freq_number = 10, ngram_number=8, ngrams = 2, pos_filter = c("NOUN", "VERB", "ADJ", "ADV"))
}
