% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_conll-u.R
\name{fst_find_stopwords}
\alias{fst_find_stopwords}
\title{Get available Finnish stopwords lists}
\usage{
fst_find_stopwords()
}
\value{
A tibble containing the stopwords lists.
}
\description{
Returns a tibble containing available Finnish stopword lists, their contents,
and the size of the lists.
}
\examples{
fst_find_stopwords()
}
