% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare.R
\name{fst_format}
\alias{fst_format}
\title{Annotate open-ended survey responses in Finnish into CoNLL-U format}
\usage{
fst_format(data, question, id, model = "ftb", weights = NULL, add_cols = NULL)
}
\arguments{
\item{data}{A dataframe of survey responses which contains an open-ended
question.}

\item{question}{The column in the dataframe which contains the open-ended
question.}

\item{id}{The column in the dataframe which contains the ids for the
responses.}

\item{model}{A Finnish language model available for [udpipe], `"ftb"`
(default) or `"tdt"`.}

\item{weights}{Optional, the column of the dataframe which contains the
respective weights for each response.}

\item{add_cols}{Optional, a column (or columns) from the dataframe which contain
other information you'd like to retain (for instance, dimension columnns for
splitting the data for comparison plots).}
}
\value{
Dataframe of annotated text in CoNLL-U format.
}
\description{
Creates a dataframe in CoNLL-U format from a list of strings of Finnish text
using the [udpipe] package and a Finnish language model.
}
\examples{
\donttest{
i <- "fsd_id"
fst_format(data = child, question = "q7", id = i)
fst_format(data = child, question = "q7", id = i, model = "tdt")
fst_format(data = child, question = "q7", id = i, weights="paino")
fst_format(child, question = "q7", id = i, add_cols = c("bv1", "bv3"))
fst_format(child, question = "q7", id = i, add_cols = "bv1, bv3")
unlink("finnish-ftb-ud-2.5-191206.udpipe")
unlink("finnish-tdt-ud-2.5-191206.udpipe")
}
}
