% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_conll-u.R
\name{fst_import_from_csv}
\alias{fst_import_from_csv}
\title{Read in and trim Finnish text survey responses}
\usage{
fst_import_from_csv(pathname, field, delim = ",")
}
\arguments{
\item{pathname}{Path and file name for input csv file.}

\item{field}{The field in the csv which contains the relevant responses.}

\item{delim}{The separator character within the csv, defaults to `","`.}
}
\value{
A list of strings.
}
\description{
`fst_import_from_csv` returns a list containing textual survey responses for
an input field with whitespace removed and NAs treated as empty strings.
}
\examples{
lonely_data <- fst_import_from_csv(pathname = "data/daF3360.csv", delim = ";", field = "q9")
bullying_data <- fst_import_from_csv(pathname = "data/daF3134.csv", delim = ";", field = "q7")
}
