% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_pos}
\alias{fst_pos}
\title{Make POS Summary Table}
\usage{
fst_pos(data)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format, with optional additional
columns.}
}
\value{
A dataframe with a count and proportion of each UPOS tag in the data
 and the full name of the tag.
}
\description{
Creates a summary table for the input CoNLL-U data which counts the number of
words of each part-of-speech tag within the data.
}
\examples{
fst_pos(fst_child)
fst_pos(fst_dev_coop)
}
