% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_prepare_svydesign.R
\name{fst_prepare_svydesign}
\alias{fst_prepare_svydesign}
\title{Read In and format Finnish survey text responses from svydesign object}
\usage{
fst_prepare_svydesign(
  svydesign,
  question,
  id,
  model = "ftb",
  stopword_list = "nltk",
  use_weights = TRUE,
  add_cols = NULL,
  manual = FALSE,
  manual_list = ""
)
}
\arguments{
\item{svydesign}{A svydesign object which contains an open-ended question.}

\item{question}{The column in the dataframe which contains the open-ended
question.}

\item{id}{The column in the dataframe which contains the ids for the
responses.}

\item{model}{A Finnish language model available for [udpipe], `"ftb"`
(default) or `"tdt"`.}

\item{stopword_list}{A valid Finnish stopword list, default is `"nltk"`, or
`"none"`.}

\item{add_cols}{Optional, a column (or columns) from the dataframe which
contain other information you'd like to retain (for instance, dimension
columnns for splitting the data for comparison plots).}

\item{weights}{Optional, the column of the dataframe which contains the
respective weights for each response.}
}
\value{
A dataframe of Finnish text in CoNLL-U format.
}
\description{
`fst_prepare_svydesign()` produces a dataframe containing Finnish survey text
 responses in CoNLL-U format with stopwords optionally removed.
}
\examples{
\donttest{
i <- "fsd_id"
child2$paino <- as.numeric((gsub(",", ".", child$paino)))
svy_child <- svydesign(id=~1, weights= ~paino, data = child)
fst_prepare_svydesign(svy_child, question = "q7", id = i, use_weights = TRUE)

dev_coop$paino <- as.numeric((gsub(",", ".", dev_coop$paino)))
svy_dev <- svydesign(id = ~1, weights = ~paino, data =dev_coop)
fst_prepare_svydesign(svy_dev, question = "q11_2", id = i, add_cols = c('q1'))
unlink("finnish-ftb-ud-2.5-191206.udpipe")
unlink("finnish-tdt-ud-2.5-191206.udpipe")
}
}
