% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare.R
\name{fst_print_available_models}
\alias{fst_print_available_models}
\title{Find treebanks available for use}
\usage{
fst_print_available_models(search = NULL)
}
\arguments{
\item{search}{An optional string for filtering the list, name of language in
English, eg. 'estonian'}
}
\value{
List of available treebanks, filtered
}
\description{
Find treebanks available for use
}
\examples{
fst_print_available_models()
fst_print_available_models(search = "swedish")
}
