% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_conll-u.R
\name{fst_rm_stopwords}
\alias{fst_rm_stopwords}
\title{Remove Finnish stopwords from CoNLL-U dataframe}
\usage{
fst_rm_stopwords(data, stopword_list = "nltk")
}
\arguments{
\item{data}{A dataframe of Finnish text in CoNLL-U format.}

\item{stopword_list}{A valid Finnish stopword list, default is `"nltk"`.}
}
\value{
A dataframe of Finnish text in CoNLL-U format.
}
\description{
Removes stopwords from a dataframe containing Finnish survey text data which
is already in CoNLL-U format.
}
\examples{
conllu_lonely_nltk <- fst_rm_stopwords(conllu_lonely)
conllu_bullying_iso <- fst_rm_stopwords(conllu_bullying, 'stopwords-iso')
}
