% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_conll-u.R
\name{fst_save}
\alias{fst_save}
\title{Save dataframe}
\usage{
fst_save(dataframe, file_name)
}
\arguments{
\item{dataframe}{A dataframe.}

\item{file_name}{The name for the saved file.}
}
\value{
Saves file in 'processed_data' folder within working directory.
}
\description{
Saves dataframe to `processed_data` subfolder of working directory.
}
\examples{
fst_save(conllu_lonely, "lonely_processed")
fst_save(conllu_lonely_nltk, "lonely_stopwords_rm")
}
