% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_summarise}
\alias{fst_summarise}
\title{Make Summary Table}
\usage{
fst_summarise(data, desc = "All respondents")
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}

\item{desc}{A string describing respondents, default is `"All respondents"`.}
}
\value{
A dataframe with summary information for the data including reponse
 rate and word counts.
}
\description{
Creates a summary table for the input CoNLL-U data which provides the
response count and proportion, total number of words, the number of unique
words, and the number of unique lemmas.
}
\examples{
fst_summarise(conllu_dev_q11_1)
fst_summarise(conllu_dev_q11_2_nltk, "Q11_2")
}
