% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_summarise_short}
\alias{fst_summarise_short}
\title{Make Simple Summary Table}
\usage{
fst_summarise_short(data)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}
}
\value{
A dataframe with summary information on word counts for the data.
}
\description{
Creates a summary table for the input CoNLL-U data which provides the total
number of words, the number of unique words, and the number of unique lemmas.
}
\examples{
fst_summarise_short(conllu_cb_bullying_iso)
fst_summarise_short(conllu_dev_q11_2_nltk)
}
