% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lonely_data}
\alias{lonely_data}
\title{Loneliness response data}
\format{
## `lonely_data`
A list of length 27851:
\describe{
  \item{lonely_data}{response text}
}
}
\source{
<https://urn.fi/urn:nbn:fi:fsd:T-FSD3360>
}
\usage{
lonely_data
}
\description{
This data contains the responses to q9 "Miltä yksinäisyys tuntuu?
Yksinäisyyteen #' liittyy usein voimakkaita tunteita, jotka ovat erilaisia
eri #' elämäntilanteissa. Jos haluat, voit kertoa yksinäisyyskokemuksiasi
tässä. #' (Avokysymys)" in the FSD3360 Helsingin Sanomat Loneliness Survey
2014.
}
\keyword{datasets}
