% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{child}
\alias{child}
\title{Child Barometer 2016 response data}
\format{
## `child_barometer_data`
A dataframe with 414 rows and 8 columns:
\describe{
  \item{fsd_id}{FSD case id}
  \item{bv1}{Gender: 1 (Boy), 2 (Girl)}
  \item{bv3}{Major region (NUTS2): 1 (Helsinki-Uusima), 2 (Etelä-Suomi), 3
   (Länsi-Suomi), 4 (Pohjois- ja Itä-Suomi)}
  \item{bv9}{Daycare before pre-school: 1 (Yes), 2 (No)}
  \item{q3}{'Which things are you good at?' response text}
  \item{q7}{'What do you think bullying is?' response text}
  \item{q11}{'What makes you happy?' response text}
  \item{paino}{Weight}
}
}
\source{
<https://urn.fi/urn:nbn:fi:fsd:T-FSD3134>
}
\usage{
child
}
\description{
This data contains background variables and the responses to q3 "Missä
asioissa olet hyvä? (Avokysymys)", q7 "Kertoisitko, mitä sinun mielestäsi
kiusaaminen on? (Avokysymys)", and q11 "Mikä tekee sinut iloiseksi?
(Avokysymys)" in the FSD3134 Lapsibarometri 2016 dataset.
}
\keyword{datasets}
