% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_comparison_concept_network.R
\name{fst_cn_compare_plot}
\alias{fst_cn_compare_plot}
\title{Concept Network- Plot comparison Concept Network}
\usage{
fst_cn_compare_plot(
  edges,
  nodes,
  concepts,
  unique_lemmas,
  name = NULL,
  concept_colour = "indianred",
  unique_colour = "darkgreen"
)
}
\arguments{
\item{edges}{Output of `fst_cn_edges`, dataframe of 'edges' connecting two words}

\item{nodes}{Output of `fst_cn_nodes`, dataframe of relevant lemmas and their associated pagerank}

\item{concepts}{List of terms which have been searched for, separated by commas.}

\item{unique_lemmas}{List of unique lemmas, output of `fst_cn_get_unique`}

\item{name}{An optional "name" for the plot, default is `NULL` and a generic
title ('Textrank extracted keyword occurences) will be used.}

\item{concept_colour}{Colour to display concept words, default is `indianred`.}

\item{unique_colour}{Colour to display unique words, default is `darkgreen`.}
}
\value{
Plot of concept network with concept and unique words (nodes) highlighted
}
\description{
Concept Network- Plot comparison Concept Network
}
\examples{
fst_cn_compare_plot(edges = q11_1_edges, nodes = q11_1_nodes, concepts = 'elintaso, köyhä, ihminen', unique_lemma = unique_2, name = "Q11_1")
fst_cn_compare_plot(q11_2_edges, q11_2_nodes, concepts = "kehitysmaa, auttaa, pyrkiä, maa, ihminen", unique_lemma = unique_2, unique_colour = 'purple' )
}
