% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_comparison_concept_network.R
\name{fst_cn_compare_plot}
\alias{fst_cn_compare_plot}
\title{Concept Network- Plot comparison Concept Network}
\usage{
fst_cn_compare_plot(
  edges,
  nodes,
  concepts,
  unique_lemmas,
  name = NULL,
  concept_colour = "#cd1719",
  unique_colour = "#4DAF4A",
  min_edge = NULL,
  max_edge = NULL,
  min_node = NULL,
  max_node = NULL
)
}
\arguments{
\item{edges}{Output of `fst_cn_edges`, dataframe of 'edges' connecting two words}

\item{nodes}{Output of `fst_cn_nodes`, dataframe of relevant lemmas and their associated pagerank}

\item{concepts}{List of terms which have been searched for, separated by commas.}

\item{unique_lemmas}{List of unique lemmas, output of `fst_cn_get_unique`}

\item{name}{An optional "name" for the plot, default is `NULL` and a generic
title ('Textrank extracted keyword occurences) will be used.}

\item{concept_colour}{Colour to display concept words, default is `indianred`.}

\item{unique_colour}{Colour to display unique words, default is `darkgreen`.}

\item{min_edge}{A numeric value for the scale of the edges, the smallest
co_occurence value for an edge across all Networks to be plotted together.}

\item{max_edge}{A numeric value for the scale of the edges, the largest
co_occurence value for an edge across all Networks to be plotted together.}

\item{min_node}{A numeric value for the scale of the nodes, the smallest
pagerank value for a node across all Networks to be plotted together.}

\item{max_node}{A numeric value for the scale of the nodes, the largest
pagerank value for a node across all Networks to be plotted together.}
}
\value{
Plot of concept network with concept and unique words (nodes) highlighted
}
\description{
Concept Network- Plot comparison Concept Network
}
\examples{
c1 <- "elintaso, köyhä, ihminen"
data1 <- conllu_dev_q11_1_nltk
e1 <- fst_cn_edges(data = data1, concepts = c1, threshold = 3)
n1 <- fst_cn_nodes(conllu_dev_q11_1_nltk, e1)

c2 <- "kehitysmaa, auttaa, pyrkiä, maa, ihminen"
e2 <- fst_cn_edges(conllu_dev_q11_2_nltk, c2, threshold = 5)
n2 <- fst_cn_nodes(conllu_dev_q11_2_nltk, e2)

c3 <- "köyhyys, nälänhätä, sota, ilmastonmuutos, puute"
e3 <- fst_cn_edges(conllu_dev_q11_3_nltk, c3, threshold = 2)
n3 <- fst_cn_nodes(conllu_dev_q11_3_nltk, e3)

u1 <- fst_cn_get_unique(n1, n2)
u2 <- fst_cn_get_unique(n1, n2, n3)

fst_cn_compare_plot(edges = e1, nodes = n1, concepts = c1, unique_lemma = u1)
fst_cn_compare_plot(e2, n2, c2, unique_lemmas = u2, unique_colour = "purple")
}
