% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_concept_network.R
\name{fst_cn_edges}
\alias{fst_cn_edges}
\title{Concept Network - Get Textrank Edges}
\usage{
fst_cn_edges(
  data,
  concept,
  threshold = NULL,
  relevant_pos = c("NOUN", "VERB", "ADJ", "ADV")
)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}

\item{concept}{List of terms to search for, separated by commas.}

\item{threshold}{A minimum number of occurrences threshold for 'edge' between
searched term and other word, default is NULL.}

\item{relevant_pos}{List of UPOS tags for inclusion, default is c("NOUN",
"VERB", "ADJ", "ADV").}
}
\value{
Dataframe of 'edges' between two connected words.
}
\description{
This function takes a string of terms (separated by commas) or a single term
and, using `fst_cn_search` find words conected to these searched terms. Then,
a dataframe is returned of 'edges' between two words which are connected
together in an frequently-occurring n-gram containing a concept term
}
\examples{
lonely_edges_2 <- fst_cn_edges(conllu_lonely, concept = 'yksinäisyys, tunne, tuntea', threshold=5)
lonely_edges <- fst_cn_edges(conllu_lonely_nltk, 'yksinäisyys, tunne, tuntea', threshold = 3)
bullying_edges <-fst_cn_edges(conllu_bullying_iso, 'kiusata, lyöminen')
}
