% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_concept_network.R
\name{fst_cn_edges}
\alias{fst_cn_edges}
\title{Concept Network - Get textrank edges}
\usage{
fst_cn_edges(data, concepts, threshold = NULL, pos_filter = NULL)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}

\item{concepts}{List of terms to search for, separated by commas.}

\item{threshold}{A minimum number of occurrences threshold for 'edge' between
searched term and other word, default is `NULL`.}

\item{pos_filter}{List of UPOS tags for inclusion, default is `NULL` to
include all UPOS tags.}
}
\value{
Dataframe of 'edges' between two connected words.
}
\description{
This function takes a string of terms (separated by commas) or a single term
and, using `fst_cn_search` find words conected to these searched terms. Then,
a dataframe is returned of 'edges' between two words which are connected
together in an frequently-occurring n-gram containing a concept term
}
\examples{
q11_1_edges <- fst_cn_edges(conllu_dev_q11_1_nltk, 'elintaso, köyhä, ihminen', threshold = 3)
q11_2_edges <- fst_cn_edges(conllu_dev_q11_2_nltk, "kehitysmaa, auttaa, pyrkiä, maa, ihminen", threshold = 5)
q11_3_edges <- fst_cn_edges(conllu_dev_q11_3_nltk, "köyhyys, nälänhätä, sota, ilmastonmuutos, puute", threshold = 2)
bullying_edges <-fst_cn_edges(conllu_bullying_iso, 'kiusata, lyöminen', pos_filter = c("NOUN", "VERB", "ADJ", "ADV"))
}
