% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_comparison_concept_network.R
\name{fst_cn_get_unique}
\alias{fst_cn_get_unique}
\title{Concept Network- Get Unique Nodes}
\usage{
fst_cn_get_unique(table1, table2, ...)
}
\arguments{
\item{table1}{The first table}

\item{table2}{The second table}

\item{...}{Any other tables you want to include}
}
\value{
Dataframe of words and whether word is unique or not
}
\description{
Takes at least two tables of nodes and pagerank (output of `fst_cn_nodes`)
and finds nodes unique to one table.
}
\examples{
unique <- fst_cn_get_unique(female_nodes, male_nodes, na_nodes)
unique_2 <- fst_cn_get_unique(q11_1_nodes, q11_2_nodes)
unique_3 <- fst_cn_get_unique(q11_1_nodes, q11_2_nodes, q11_3_nodes)
}
