% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_comparison_concept_network.R
\name{fst_cn_get_unique}
\alias{fst_cn_get_unique}
\title{Concept Network- Get Unique Nodes}
\usage{
fst_cn_get_unique(table1, table2, ...)
}
\arguments{
\item{table1}{The first table}

\item{table2}{The second table}

\item{...}{Any other tables you want to include}
}
\value{
Dataframe of words and whether word is unique or not
}
\description{
Takes at least two tables of nodes and pagerank (output of `fst_cn_nodes`)
and finds nodes unique to one table.
}
\examples{
con1 <- "elintaso, köyhä, ihminen"
data1 <- conllu_dev_q11_1_nltk
e1 <- fst_cn_edges(data = data1, concepts = con1, threshold = 3)
q11_1_nodes <- fst_cn_nodes(conllu_dev_q11_1_nltk, e1)

con2 <- "kehitysmaa, auttaa, pyrkiä, maa, ihminen"
e2 <- fst_cn_edges(conllu_dev_q11_2_nltk, con2, threshold = 5)
q11_2_nodes <- fst_cn_nodes(conllu_dev_q11_2_nltk, e2)

con3 <- "köyhyys, nälänhätä, sota, ilmastonmuutos, puute"
e3 <- fst_cn_edges(conllu_dev_q11_3_nltk, con3, threshold = 2)
q11_3_nodes <- fst_cn_nodes(conllu_dev_q11_3_nltk, e3)

fst_cn_get_unique(q11_1_nodes, q11_2_nodes)
fst_cn_get_unique(q11_1_nodes, q11_2_nodes, q11_3_nodes)
}
