% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_concept_network.R
\name{fst_cn_nodes}
\alias{fst_cn_nodes}
\title{Concept Network - Get textrank nodes}
\usage{
fst_cn_nodes(data, edges, pos_filter = NULL)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}

\item{edges}{Output of fst_cn_edges, dataframe of 'edges' connecting two words}

\item{pos_filter}{List of UPOS tags for inclusion, default is `NULL` to
include all UPOS tags.}
}
\value{
A dataframe containing relevant lemmas and their associated pagerank
}
\description{
Concept Network - Get textrank nodes
}
\examples{
con1 <- "elintaso, köyhä, ihminen"
data1 <- conllu_dev_q11_1_nltk
e1 <- fst_cn_edges(data = data1, concepts = con1, threshold = 3)
fst_cn_nodes(conllu_dev_q11_1_nltk, e1)

con2 <- "kehitysmaa, auttaa, pyrkiä, maa, ihminen"
e2 <- fst_cn_edges(conllu_dev_q11_2_nltk, con2, threshold = 5)
fst_cn_nodes(conllu_dev_q11_2_nltk, e2)

con3 <- "köyhyys, nälänhätä, sota, ilmastonmuutos, puute"
e3 <- fst_cn_edges(conllu_dev_q11_3_nltk, con3, threshold = 2)
fst_cn_nodes(conllu_dev_q11_3_nltk, e3)

con4 <- "kiusata, lyöminen"
cb <- conllu_cb_bullying_iso
e4 <- fst_cn_edges(cb, con4, pos_filter = c("NOUN", "VERB", "ADJ", "ADV"))
fst_cn_nodes(conllu_cb_bullying_iso, e4, c("NOUN", "VERB", "ADJ", "ADV"))
}
