% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_concept_network.R
\name{fst_cn_plot}
\alias{fst_cn_plot}
\title{Plot Concept Network}
\usage{
fst_cn_plot(edges, nodes, concepts, title = NULL)
}
\arguments{
\item{edges}{Output of `fst_cn_edges`, dataframe of 'edges' connecting two
words}

\item{nodes}{Output of `fst_cn_nodes`, dataframe of relevant lemmas and their
associated pagerank}

\item{concepts}{List of terms which have been searched for, separated by
commas.}

\item{title}{Optional title for plot, default is `NULL` and a generic title
('Textrank extracted keyword occurrences) will be used.}
}
\value{
Plot of Concept Network
}
\description{
Plot Concept Network
}
\examples{
fst_cn_plot(edges = q11_1_edges, nodes = q11_1_nodes, concepts = "elintaso, köyhä, ihminen")
fst_cn_plot(edges = q11_2_edges, nodes = q11_2_nodes, concepts = "kehitysmaa, auttaa, pyrkiä, maa, ihminen")
fst_cn_plot(edges = q11_3_edges, nodes = q11_3_nodes, concepts = "köyhyys, nälänhätä, sota, ilmastonmuutos, puute")
fst_cn_plot(edges = bullying_edges, nodes = bullying_nodes, concepts = "kiusata, lyöminen")
}
