% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_concept_network.R
\name{fst_cn_search}
\alias{fst_cn_search}
\title{Concept Network - Search Textrank for Concepts}
\usage{
fst_cn_search(data, concept, relevant_pos = c("NOUN", "VERB", "ADJ", "ADV"))
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}

\item{concept}{String of terms to search for, separated by commas.}

\item{relevant_pos}{List of UPOS tags for inclusion, default is c("NOUN",
"VERB", "ADJ", "ADV").}
}
\value{
Dataframe of n-grams containing searched terms
}
\description{
This function takes a string of terms (separated by commas) or a single term
and, using `textrank_keywords` from `textrank` package, filters data based on
`relevant_pos` and finds words connected to search terms.
}
\examples{
q11_1_concepts <- fst_cn_search(conllu_dev_q11_1_nltk, "elintaso, köyhä, ihminen", relevant_pos = "NOUN")
q11_2_concepts <- fst_cn_search(conllu_dev_q11_2_nltk, "kehitysmaa, auttaa, pyrkiä, maa, ihminen", relevant_pos = c("ADV", "ADJ"))
q11_3_concepts <- fst_cn_search(conllu_dev_q11_3_nltk, "köyhyys, nälänhätä, sota, ilmastonmuutos, puute")
bullying_concepts <- fst_cn_search(conllu_bullying_iso, 'kiusata, lyöminen, lyödä, potkia')
}
