% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_concept_network.R
\name{fst_cn_search}
\alias{fst_cn_search}
\title{Concept Network - Search textrank for concepts}
\usage{
fst_cn_search(data, concepts, pos_filter = NULL)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}

\item{concepts}{String of terms to search for, separated by commas.}

\item{pos_filter}{List of UPOS tags for inclusion, default is `NULL` to
include all UPOS tags.}
}
\value{
Dataframe of n-grams containing searched terms
}
\description{
This function takes a string of terms (separated by commas) or a single term
and, using `textrank_keywords` from `textrank` package, filters data based on
`pos_filter` and finds words connected to search terms.
}
\examples{
con1 <- "elintaso, köyhä, ihminen"
fst_cn_search(conllu_dev_q11_1_nltk, con1)

con2 <- "kehitysmaa, auttaa, pyrkiä, maa, ihminen"
pf1 <- c("NOUN", "ADV", "ADJ")
fst_cn_search(conllu_dev_q11_2_nltk, con2, pos_filter = pf1)

con3 <- "köyhyys, nälänhätä, sota, ilmastonmuutos, puute"
fst_cn_search(conllu_dev_q11_3_nltk, con3, pos_filter = "NOUN")

con4 <- "kiusata, lyöminen, lyödä, potkia"
pf2 <- c("NOUN", "VERB", "ADJ", "ADV")
fst_cn_search(conllu_cb_bullying_iso, con4, pos_filter = pf2)
}
