% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_comparison_functions.R
\name{fst_comparison_cloud}
\alias{fst_comparison_cloud}
\title{Make comparison cloud}
\usage{
fst_comparison_cloud(
  data,
  field,
  pos_filter = NULL,
  max = 100,
  use_svydesign_weights = FALSE,
  id = "",
  svydesign = NULL,
  use_column_weights = FALSE,
  exclude_nulls = FALSE,
  rename_nulls = "null_data"
)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format with additional `field`
column for splitting data.}

\item{field}{Column in `data` used for splitting groups}

\item{pos_filter}{List of UPOS tags for inclusion, default is `NULL` which
means all word types included.}

\item{max}{The maximum number of words to display, default is `100`.}

\item{exclude_nulls}{Whether to include NULLs in `field` column, default is
`FALSE`}

\item{rename_nulls}{What to fill NULL values with if `exclude_nulls = FALSE`.}
}
\value{
A comparison cloud from wordcloud package.
}
\description{
Creates a comparison wordcloud showing words that occur differently between
each group. Data is split based on different values in the `field` column of
formatted data. Results will be shown within the plots pane.
}
\examples{
fst_comparison_cloud(fst_child, 'bv1', max = 50)
fst_child_3 <- within(fst_child, rm(weight))
child$paino <- as.numeric((gsub(",", ".", child$paino)))
s <- svydesign(id=~1, weights= ~paino, data = child)
i <- 'fsd_id'
fst_comparison_cloud(fst_child_3, 'bv1', use_svydesign_weights = TRUE, id = i, svydesign = s)
fst_comparison_cloud(fst_dev_coop, 'q3', use_column_weights = TRUE)
}
