% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_concept_network.R
\name{fst_concept_network}
\alias{fst_concept_network}
\title{Concept Network - Make Concept Network Plot}
\usage{
fst_concept_network(
  data,
  concepts,
  threshold = NULL,
  relevant_pos = c("NOUN", "VERB", "ADJ", "ADV")
)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}

\item{threshold}{A minimum number of occurrences threshold for 'edge' between
searched term and other word, default is NULL.}

\item{relevant_pos}{List of UPOS tags for inclusion, default is c("NOUN",
"VERB", "ADJ", "ADV").}

\item{input_word}{List of terms to search for, separated by commas.}
}
\value{
Plot of concept network
}
\description{
Concept Network - Make Concept Network Plot
}
\examples{
fst_concept_network(conllu_lonely_nltk, concepts = "yksinäisyys, tunne, tuntea", threshold=3)
fst_concept_network(conllu_lonely_nltk, concepts = "tunne, tuntea", threshold=3)
fst_concept_network(conllu_lonely_nltk, concepts = "tunne, tuntea")
fst_concept_network(conllu_lonely, concepts = "yksinäisyys, tunne, tuntea", threshold=5)
fst_concept_network(conllu_bullying_iso, concepts = 'kiusata, lyöminen')
}
