% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_freq}
\alias{fst_freq}
\title{Make Top Words Table}
\usage{
fst_freq(data, number = 20, pos_filter = NULL)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}

\item{number}{The number of top words to return, default is `20`.}

\item{pos_filter}{List of UPOS tags for inclusion, default is `NULL` which
means all word types included}
}
\value{
A table of the most frequent words (unigrams) in the data
}
\description{
Creates a table of the most frequently-occurring words (unigrams) within the
data.
}
\examples{
fst_freq(conllu_bullying_iso)
fst_freq(conllu_bullying_iso, 10, c("NOUN", "VERB", "ADJ", "ADV"))
fst_freq(conllu_dev_q11_1_nltk, 15, "NOUN")
fst_freq(conllu_dev_q11_1_snow, 15, "ADV")
}
