% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_comparison_functions.R
\name{fst_freq_compare}
\alias{fst_freq_compare}
\title{Compare and plot top words}
\usage{
fst_freq_compare(
  data1,
  data2,
  data3 = NULL,
  data4 = NULL,
  number = 10,
  pos_filter = NULL,
  name1 = "Group 1",
  name2 = "Group 2",
  name3 = "Group 3",
  name4 = "Group 4",
  unique_colour = "indianred",
  strict = TRUE
)
}
\arguments{
\item{data1}{A dataframe of text in CoNLL-U format for the first plot.}

\item{data2}{A dataframe of text in CoNLL-U format for the second plot.}

\item{data3}{An optional dataframe of text in CoNLL-U format for the third
plot, default is `NULL`}

\item{data4}{An optional dataframe of text in CoNLL-U format for the fourth
plot, default is `NULL`}

\item{number}{The number of top words to return, default is `10`.}

\item{pos_filter}{List of UPOS tags for inclusion, default is `NULL` which
means all word types included}

\item{name1}{An optional "name" for the first plot, default is `NULL`}

\item{name2}{An optional "name" for the second plot, default is `NULL`}

\item{name3}{An optional "name" for the third plot, default is `NULL`}

\item{name4}{An optional "name" for the fourth plot, default is `NULL`}

\item{unique_colour}{Colour to display unique words, default is `indianred`.}

\item{strict}{Whether to strictly cut-off at `number` (ties are
alphabetically ordered), default is `TRUE`}
}
\value{
Between 2 and 4 plots of Top n-grams in the plots pane with unique
n-grams highlighted
}
\description{
Find top and unique top words for between 2 and 4 sets of prepared data.
Results will be shown within the plots pane. If 2 or 3 plots, they will be in
a single row, if there are 4 plots, they will be in 2 rows of 2.
}
\examples{
fst_freq_compare(conllu_dev_q11_1_m_nltk, conllu_dev_q11_1_f_nltk, number = 10)
fst_freq_compare(conllu_dev_q11_1_m_nltk, conllu_dev_q11_1_f_nltk, conllu_dev_q11_1_na_nltk, number = 5)
fst_freq_compare(conllu_dev_q11_1_m_nltk, conllu_dev_q11_1_f_nltk, conllu_dev_q11_1_na_nltk,number = 15, unique_colour = 'pink', pos_filter = c("NOUN", "VERB", "ADJ", "ADV"), name1 = 'Male', name2 = 'Female', name3 = 'Not Specified')
fst_freq_compare(conllu_dev_q11_1_m_nltk, conllu_dev_q11_1_f_nltk, conllu_dev_q11_1_na_nltk, number = 10, name1 = 'Female', name2 = 'Male', name3 = 'Gender Not Specified')
fst_freq_compare(conllu_dev_q11_1_m_nltk, conllu_dev_q11_1_f_nltk, conllu_dev_q11_1_na_nltk, number = 10, name1 = 'Female', name2 = 'Male', name3 = 'Gender Not Specified', strict = FALSE)
}
