% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_comparison_functions.R
\name{fst_freq_compare}
\alias{fst_freq_compare}
\title{Compare and plot top words}
\usage{
fst_freq_compare(
  data,
  field,
  number = 10,
  norm = NULL,
  pos_filter = NULL,
  strict = TRUE,
  use_svydesign_weights = FALSE,
  id = "",
  svydesign = NULL,
  use_column_weights = FALSE,
  exclude_nulls = FALSE,
  rename_nulls = "null_data",
  unique_colour = "indianred"
)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format with additional `field`
column for splitting data.}

\item{field}{Column in `data` used for splitting groups}

\item{number}{The number of n-grams to return, default is `10`.}

\item{norm}{The method for normalising the data. Valid settings are
`"number_words"` (the number of words in the responses), `"number_resp"`
(the number of responses), or `NULL` (raw count returned, default, also used
when weights are applied).}

\item{pos_filter}{List of UPOS tags for inclusion, default is `NULL` which
means all word types included.}

\item{strict}{Whether to strictly cut-off at `number` (ties are
alphabetically ordered), default is `TRUE`.}

\item{use_svydesign_weights}{Option to weight words in the wordcloud using
weights from  a svydesign object containing the raw data, default is `FALSE`}

\item{id}{ID column from raw data, required if `use_svydesign_weights = TRUE`
and must match the `docid` in formatted `data`.}

\item{svydesign}{A svydesign object which contains the raw data and weights.}

\item{use_column_weights}{Option to weight words in the wordcloud using
weights from  formatted data which includes addition `weight` column,
default is `FALSE`}

\item{exclude_nulls}{Whether to include NULLs in `field` column, default is
`FALSE`}

\item{rename_nulls}{What to fill NULL values with if `exclude_nulls = FALSE`.}

\item{unique_colour}{Colour to display unique words, default is
`"indianred"`.}

\item{data4}{An optional dataframe of text in CoNLL-U format for the fourth}
}
\value{
Plots of most frequent words in the plots pane with unique words
 highlighted.
}
\description{
Find top and unique top words for different groups of participants. Data is
split based on different values in the `field` column of formatted data.
Results will be shown within the plots pane.
}
\examples{
fst_freq_compare(fst_child, 'bv1', number = 10, norm = "number_resp")
fst_freq_compare(fst_child, 'bv1', number = 10, norm = NULL)
fst_child_3 <- within(fst_child, rm(weight))
fst_freq_compare(fst_child_3, 'bv1', number = 10, use_svydesign_weights = TRUE, id = 'fsd_id', svydesign = svy_child)
fst_freq_compare(fst_child, 'bv1', number = 10, use_column_weights = TRUE, strict = FALSE, unique_colour = 'purple')
}
