% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_freq_plot}
\alias{fst_freq_plot}
\title{Make Top Words plot}
\usage{
fst_freq_plot(table, number = NULL, name = NULL)
}
\arguments{
\item{table}{Output of `fst_freq_table()` or `fst_ngrams_table()`.}

\item{number}{Optional number of n-grams for the title, default is `NULL`.}

\item{name}{An optional "name" for the plot to add to title, default is
`NULL`.}
}
\value{
Plot of top words.
}
\description{
Plots most common words.
}
\examples{
pf <- c("NOUN", "VERB", "ADJ", "ADV")
top_words <- fst_freq_table(fst_child, number = 15, pos_filter = pf)
fst_freq_plot(top_words, number = 15, name = "Bullying")
}
