% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_freq_plot}
\alias{fst_freq_plot}
\title{Make Top Words plot}
\usage{
fst_freq_plot(table, number = NULL, name = NULL)
}
\arguments{
\item{table}{Output of `fst_get_top_words` or `fst_get_top_ngrams`}

\item{number}{The number of n-grams, default is `10`.}

\item{name}{An optional "name" for the plot to add to title, default is `NULL`}
}
\value{
Plot of top words.
}
\description{
Plots most common words.
}
\examples{
fst_freq_plot(top_f, number = 15, name = 'Female')
fst_freq_plot(top_m)
fst_freq_plot(top_na, number = 15)
}
