% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_get_top_ngrams}
\alias{fst_get_top_ngrams}
\title{Make Top N-grams Table}
\usage{
fst_get_top_ngrams(
  data,
  number = 10,
  ngrams = 1,
  pos_filter = NULL,
  strict = TRUE
)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}

\item{number}{The number of n-grams to return, default is `10`.}

\item{ngrams}{The type of n-grams to return, default is `1`.}

\item{pos_filter}{List of UPOS tags for inclusion, default is `NULL` which
means all word types included}

\item{strict}{Whether to strictly cut-off at `number` (ties are
alphabetically ordered), default is `TRUE`}
}
\value{
A table of the most frequently occurring n-grams in the data.
}
\description{
Creates a table of the most frequently-occurring ngrams within the
data.
}
\examples{
fst_get_top_ngrams(conllu_dev_q11_1_nltk)
fst_get_top_ngrams(conllu_dev_q11_1_nltk, number = 10, ngrams = 1)
top_bullying_ngrams <- fst_get_top_ngrams(conllu_bullying, number = 15, pos_filter = c("NOUN", "VERB", "ADJ", "ADV"))
topn_f <- fst_get_top_ngrams(conllu_f_nltk, number = 15, ngrams = 2)
topn_m <- fst_get_top_ngrams(conllu_m_nltk, number = 15, ngrams = 2)
topn_na <- fst_get_top_ngrams(conllu_na_nltk, number = 15, ngrams = 2)
}
