% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_get_top_words}
\alias{fst_get_top_words}
\title{Make Top Words Table}
\usage{
fst_get_top_words(data, number = 10, pos_filter = NULL, strict = TRUE)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}

\item{number}{The number of top words to return, default is `10`.}

\item{pos_filter}{List of UPOS tags for inclusion, default is `NULL` which
means all word types included}

\item{strict}{Whether to strictly cut-off at `number` (ties are
alphabetically ordered), default is `TRUE`}
}
\value{
A table of the most frequently occurring words in the data
}
\description{
Creates a table of the most frequently-occurring words (unigrams) within the
data.
}
\examples{
fst_get_top_words(conllu_dev_q11_1_nltk, number = 15)
fst_get_top_words(conllu_dev_q11_1_nltk, number = 15, strict = FALSE)
top_bullying_words <- fst_get_top_words(conllu_bullying, number = 15, pos_filter = c("NOUN", "VERB", "ADJ", "ADV"))
top_f <- fst_get_top_words(conllu_f_nltk, number = 15)
top_m <- fst_get_top_words(conllu_m_nltk, number = 15)
top_na <- fst_get_top_words(conllu_na_nltk, number = 15)
}
