% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_comparison_functions.R
\name{fst_join_unique}
\alias{fst_join_unique}
\title{Merge Nn-grams table with unique words}
\usage{
fst_join_unique(table, unique_table)
}
\arguments{
\item{table}{Output of `fst_get_top_words` or `fst_get_top_ngrams`}

\item{unique_table}{Output of `fst_get_unique_ngrams`}
}
\value{
A table of top n-grams, frequency, and whether the n-gram is "unique"
}
\description{
Merges list of unique words from `fst_get_unique_ngrams` with output of
`fst_get_top_ngrams` or `fst_get_top_words` so that unique words can be
displayed on comparison plots.
}
\examples{
top_fu <- fst_join_unique(top_f, unique_w)
top_mu <- fst_join_unique(top_m, unique_w)
top_nau <- fst_join_unique(top_na, unique_w)
topn_fu <- fst_join_unique(topn_f, unique_n)
topn_mu <- fst_join_unique(topn_m, unique_n)
topn_nau <- fst_join_unique(topn_na, unique_n)
}
