% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_length_summary}
\alias{fst_length_summary}
\title{Make Length Summary Table}
\usage{
fst_length_summary(data, desc = "All responses", incl_sentences = TRUE)
}
\arguments{
\item{data}{dataframe of text in CoNLL-U format, with optional additional
columns.}

\item{desc}{An optional string describing responses in table, default is
`"All responses"`.}

\item{incl_sentences}{Whether to include sentence data in table, default is
`TRUE`.}
}
\value{
Table summarising distribution of lengths of responses.
}
\description{
Creates a table summarising  distribution of the length of responses.
}
\examples{
fst_length_summary(fst_child, incl_sentences = FALSE)
fst_length_summary(fst_dev_coop, desc = "Q11_3")
}
