% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_length_summary}
\alias{fst_length_summary}
\title{Make Length Summary Table}
\usage{
fst_length_summary(data, desc = NULL, incl_sentences = TRUE)
}
\arguments{
\item{desc}{An optional string describing respondents, default is `NULL`}

\item{incl_sentences}{Whether to include sentence data in table, default is
`TRUE`}

\item{dataA}{dataframe of text in CoNLL-U format.}
}
\value{
Table summarising distribution of lengths of responses
}
\description{
Create table summarising length of responses
}
\examples{
fst_length_summary(conllu_dev_q11_1, incl_sentences = FALSE)
fem <- fst_length_summary(conllu_dev_q11_1, desc = 'Female')
}
