% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_ngrams}
\alias{fst_ngrams}
\title{Make Top N-Grams Table}
\usage{
fst_ngrams(data, number = 20, ngrams = 2)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}

\item{number}{The number of top n-grams to return, default is `20`.}

\item{ngrams}{The type of n-grams to return, default is `2`.}
}
\value{
A table of the most frequent n-grams in the data
}
\description{
Creates a table of the most frequently-occurring sets of words (n-grams)
within the data.
}
\examples{
fst_ngrams(conllu_bullying_iso)
fst_ngrams(conllu_dev_q11_2_nltk, number = 10, ngrams=3)
fst_ngrams(conllu_dev_q11_3_nltk, ngrams = 4)
}
