% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_comparison_functions.R
\name{fst_ngrams_compare_plot}
\alias{fst_ngrams_compare_plot}
\title{Plot comparison n-grams}
\usage{
fst_ngrams_compare_plot(
  table,
  number = 10,
  ngrams = 1,
  unique_colour = "indianred",
  name = NULL,
  override_title = NULL
)
}
\arguments{
\item{table}{The table of n-grams, output of `get_unique_ngrams`}

\item{number}{The number of n-grams, default is `10`.}

\item{ngrams}{The type of n-grams, default is `1`.}

\item{unique_colour}{Colour to display unique words, default is `indianred`.}

\item{name}{An optional "name" for the plot, default is `NULL`}
}
\value{
Plot of top n-grams with unique terms highlighted.
}
\description{
Plots frequency n-grams with unique n-grams highlighted.
}
\examples{
fst_ngrams_compare_plot(top_fu, ngrams = 1, name = 'Female')
plot_top_fu <- fst_ngrams_compare_plot(top_fu, ngrams = 1, number = 8, override_title = 'Female')
plot_top_mu <- fst_ngrams_compare_plot(top_mu, ngrams = 1, override_title = 'Male')
plot_top_nau <- fst_ngrams_compare_plot(top_nau, ngrams = 1, override_title = 'No Gender Specified')
plot_topn_fu <- fst_ngrams_compare_plot(topn_fu, ngrams = 2, name = 'Female')
plot_topn_mu <- fst_ngrams_compare_plot(topn_mu, ngrams = 2, name = 'Male')
plot_top_mu + ggplot2::ggtitle("Plot of Top Male Words") + ggplot2::xlab("Count") + ggplot2::ylab("Top Words") # How to overwrite
plot_topn_nau <- fst_ngrams_compare_plot(topn_nau, ngrams = 2, name = 'No Gender Specified')
}
