% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_comparison_functions.R
\name{fst_plot_multiple}
\alias{fst_plot_multiple}
\title{Display comparison plots}
\usage{
fst_plot_multiple(plot1, plot2, plot3 = NULL, plot4 = NULL, main_title = NULL)
}
\arguments{
\item{plot1}{First plot to display}

\item{plot2}{Second plot to display}

\item{plot3}{Optional third plot to display, defaul is `NULL`}

\item{plot4}{Optional fourth plot to display, defaul is `NULL`}
}
\value{
Up to 4 plots within the plots pane.
}
\description{
Display between 2 and 4 plots within the plots pane. If 2 or 3 plots, they
will be in a single row, if there are 4 plots, they will be in 2 rows of 2.
}
\examples{
fst_plot_multiple(plot_top_fu, plot_top_mu, plot_top_nau, main_title = "Comparison Plots")
fst_plot_multiple(plot_topn_fu, plot_topn_mu, plot_topn_nau, plot_topn_nau)
fst_plot_multiple(plot_topn_fu, plot_topn_mu, main_title = "Comparison Plots")
}
