% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_pos}
\alias{fst_pos}
\title{Make POS Summary Table}
\usage{
fst_pos(data)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}
}
\value{
A dataframe with a count of each UPOS tag in the data and the full
 name of the tag
}
\description{
Creates a summary table for the input CoNLL-U data which counts the number of
words of each part-of-speech tag within the data.
}
\examples{
bullying_pos_table <- fst_pos(conllu_bullying_iso)
lonely_pos_table <- fst_pos(conllu_lonely_nltk)
}
