% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_comparison_functions.R
\name{fst_pos_compare}
\alias{fst_pos_compare}
\title{Compare parts-of-speech}
\usage{
fst_pos_compare(
  data1,
  data2,
  data3 = NULL,
  data4 = NULL,
  name1 = "Group 1",
  name2 = "Group 2",
  name3 = "Group 3",
  name4 = "Group 4"
)
}
\arguments{
\item{data1}{A dataframe of text in CoNLL-U format for the first group}

\item{data2}{A dataframe of text in CoNLL-U format for the second group.}

\item{data3}{An optional dataframe of text in CoNLL-U format for the third
group, default is `NULL`}

\item{data4}{An optional dataframe of text in CoNLL-U format for the fourth
group, default is `NULL`}

\item{name1}{An optional "name" for the first group, default is `NULL`}

\item{name2}{An optional "name" for the second group, default is `NULL`}

\item{name3}{An optional "name" for the third group, default is `NULL`}

\item{name4}{An optional "name" for the fourth group, default is `NULL`}
}
\value{
Table of POS tag counts for the groups
}
\description{
Compare words in responses based on part-of-speech tagging for between 2 and
4 sets of prepared data.
}
\examples{
fst_pos_compare(conllu_dev_q11_1_f, conllu_dev_q11_1_m, conllu_dev_q11_1_na, conllu_dev_q11_1, "Female", "Male", "No Gender Specified", "All")
fst_pos_compare(conllu_dev_q11_1_f, conllu_dev_q11_1_m, name1 = "Female", name2 = "Male")
pos_table <- fst_pos_compare(data1 = conllu_dev_q11_1, name1 = "All", data2 = conllu_dev_q11_1_m, name2 = "Male", data3 = conllu_dev_q11_1_f, name3 = "Female", data4 = conllu_dev_q11_1_na, name4 = "Not Spec")
}
