% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare.R
\name{fst_prepare}
\alias{fst_prepare}
\title{Read In and format Finnish survey text responses}
\usage{
fst_prepare(
  data,
  question,
  id,
  model = "ftb",
  stopword_list = "nltk",
  weights = NULL,
  add_cols = NULL,
  manual = FALSE,
  manual_list = ""
)
}
\arguments{
\item{data}{A dataframe of survey responses which contains an open-ended
question.}

\item{question}{The column in the dataframe which contains the open-ended
question.}

\item{id}{The column in the dataframe which contains the ids for the
responses.}

\item{model}{A Finnish language model available for [udpipe], `"ftb"`
(default) or `"tdt"`.}

\item{stopword_list}{A valid Finnish stopword list, default is `"nltk"`, or
`"none"`.}

\item{weights}{Optional, the column of the dataframe which contains the
respective weights for each response.}

\item{add_cols}{Optional, a column (or columns) from the dataframe which
contain other information you'd like to retain (for instance, dimension
columnns for splitting the data for comparison plots).}
}
\value{
A dataframe of Finnish text in CoNLL-U format.
}
\description{
`fst_prepare_conllu()` produces a dataframe containing Finnish survey text
 responses in CoNLL-U format with stopwords optionally removed.
}
\examples{
\donttest{
i <- "fsd_id"
cb <- child
dev <- dev_coop
fst_prepare(data = cb, question = "q7", id = 'fsd_id', weights = 'paino')
fst_prepare(data = dev, question = "q11_2", id = i, add_cols = c('q1'))
unlink("finnish-ftb-ud-2.5-191206.udpipe")
unlink("finnish-tdt-ud-2.5-191206.udpipe")
}
}
