% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_conll-u.R
\name{fst_prepare_conllu}
\alias{fst_prepare_conllu}
\title{Read In and format Finnish survey text responses}
\usage{
fst_prepare_conllu(data, field, model = "ftb", stopword_list = "nltk")
}
\arguments{
\item{data}{A dataframe of survey responses which contains an open-ended question.}

\item{field}{The field in the dataframe which contains the open-ended question.}

\item{model}{A Finnish language model available for [udpipe], `ftb` (default)
or `tdt`.}

\item{stopword_list}{A valid Finnish stopword list, default is `nltk`, or
"none".}
}
\value{
CoNLL-U dataframe
}
\description{
`fst_prepare_conllu` produces a dataframe (and saves as csv) containing
Finnish survey text reponses in CoNLL-U format with stopwords removed.
}
\examples{
prepd_bullying <- fst_prepare_conllu(data = bullying_data, field = "q7", stopword_list = "stopwords-iso")
prepd_dev_q11_2 <- fst_prepare_conllu(data = dev_data, field = "q11_2", stopword_list = "snowball")
}
