% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_conll-u.R
\name{fst_prepare_conllu}
\alias{fst_prepare_conllu}
\title{Read In and format Finnish survey text responses}
\usage{
fst_prepare_conllu(
  pathname,
  field,
  file_name,
  delim = ",",
  model = "ftb",
  stopword_list = "nltk"
)
}
\arguments{
\item{pathname}{Path and file name for input csv file.}

\item{field}{The field in the csv which contains the relevant responses.}

\item{file_name}{The name for the saved file.}

\item{delim}{The separator character within the csv, defaults to `","`.}

\item{model}{A Finnish language model available for [udpipe], `ftb` (default)
or `tdt`.}

\item{stopword_list}{A valid Finnish stopword list, default is `nltk`, or
"none".}
}
\value{
Saves file in 'processed_data' folder within working directory.
}
\description{
`fst_prepare_conllu` produces a dataframe (and saves as csv) containing
Finnish survey text reponses in CoNLL-U format with stopwords removed.
}
\examples{
prepd_bullying <- fst_prepare_conllu(pathname = "data/daF3134.csv", field = "q7", stopword_list = "nltk", file_name = "bullying", delim = ";")
prepd_lonely <- fst_prepare_conllu("data/daF3360.csv", field = "q9", stopword_list = "none", file_name="lonely", delim = ";")
}
