% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_summarise}
\alias{fst_summarise}
\title{Make Summary Table}
\usage{
fst_summarise(data, desc = "All responses")
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format, with optional additional
columns.}

\item{desc}{A string describing responses in table, default is `"All
responses"`.}
}
\value{
A dataframe with summary information for the data including response
 rate and word counts.
}
\description{
Creates a summary table for the input CoNLL-U data which provides the
response count and proportion, total number of words, the number of unique
words, and the number of unique lemmas.
}
\examples{
fst_summarise(fst_child)
fst_summarise(fst_dev_coop, "Q11_3")
}
