% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_summarise}
\alias{fst_summarise}
\title{Make Summary Table}
\usage{
fst_summarise(data)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}
}
\value{
A dataframe with summary info for the data.
}
\description{
Creates a summary table for the input CoNLL-U data which provides the total
number of words, the number of unique words, and the number of unique lemmas.
}
\examples{
bullying_summary <- fst_summarise(conllu_bullying_iso)
lonely_summary <- fst_summarise(conllu_lonely_nltk)
}
