% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_use_svydesign}
\alias{fst_use_svydesign}
\title{Add svydeesign weights to CoNLL-U data}
\usage{
fst_use_svydesign(data, svydesign, id)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format}

\item{svydesign}{A svydesign object containing the raw data which produced
the `data`}

\item{id}{ID column from raw data, must match the `docid` in formatted `data`}
}
\value{
A dataframe of text in CoNLL-U format plus a `'weight'` column
}
\description{
This function takes data in CoNLL-U format and a svydesign (from survey
packge) object with weights in it and merges the weights into the formatted
data.
}
\examples{
child$paino <- as.numeric((gsub(",", ".", child$paino)))
svy_child <- svydesign(id=~1, weights= ~paino, data = child)
fst_use_svydesign(data = fst_child_2, svydesign = svy_child, id = 'fsd_id')

dev_coop$paino <- as.numeric((gsub(",", ".", dev_coop$paino)))
svy_dev <- svydesign(id = ~1, weights = ~paino, data = dev_coop)
fst_use_svydesign(data = fst_dev_coop_2, svydesign = svy_dev, id = 'fsd_id')
}
