% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_exploration.R
\name{fst_wordcloud}
\alias{fst_wordcloud}
\title{Make Wordcloud}
\usage{
fst_wordcloud(data, pos_filter = NULL)
}
\arguments{
\item{data}{A dataframe of text in CoNLL-U format.}

\item{pos_filter}{List of UPOS tags for inclusion, default is `NULL` which
means all word types included.}
}
\value{
A wordcloud from the data
}
\description{
Creates a wordcloud from CoNLL-U data of frequently-occuring words.
}
\examples{
fst_wordcloud(conllu_bullying_iso)
fst_wordcloud(conllu_bullying_iso, pos_filter = c("NOUN", "VERB", "ADJ", "ADV"))
fst_wordcloud(conllu_dev_q11_1_snow, pos_filter = "VERB")
}
